[Refactor] name package same as content of the package
