+++ PokeCards/app/src/main/res/layout/card_details_activity.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.themanol.pokecards.details.viewmodels.CardDetailsViewModel" />
+            type="com.themanol.pokecards.card_details.viewmodels.CardDetailsViewModel" />
     </data>
 
     <RelativeLayout
@@ -18,7 +18,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context="com.themanol.pokecards.details.CardDetailsActivity">
+        tools:context="com.themanol.pokecards.card_details.CardDetailsActivity">
 
 
         <TextView
